backend/requirements.txt
uvicorn
gunicorn
fastapi==0.115.9
python-dotenv
google-generativeai==0.8.3
pydantic==2.10.3
python-multipart==0.0.20
aiofiles==23.2.1